Add extention of InputStream to pass bytes to XML parser
DocumentSplittingInputStream will help me process XML bytes sent
over the socket connection before trying to parse them with
XMLParser. Currently the client is hardcoded to send a test
XML "file", and the server will stop processing it once it sees
a closing "</stream>" tag, the signal of the end of an XMPP
stream.

TODO: In the future, XMPP stanzas will need to be processed
as they come in. Implement the DocumentSplittingInputStream so
that it gives XMLParser stanzas to parse as it receives them.