Add a reading thread on client end allowing client to read stream open message
Currently both the client and server are using XMPPReader and XMPPWriter.
The set up as of now has the server begin by listening and the client begin
by sending a stream open message and then waiting until it receives a response
from the server. Right now, both the client's and server's XMPPReader break out
of the read loop after receiving a stream open connection to allow me to test
the interaction. On the server side, I make another call to read() after the
connection is open. This should be fixed.